+++ tests/libs/estdlib/tests.erl
@@ -83,7 +83,8 @@ get_non_networking_tests(_OTPVersion) ->
         test_supervisor,
         test_lists_subtraction,
         test_os,
-        test_file
+        test_file,
+        test_filename
     ].
 
 get_networking_tests(OTPVersion) when